First pulling-out of hr-alg3
