+++ src/c5412.rs
@@ -20,10 +20,6 @@ use embassy_time::Timer;
 use core::sync::atomic::Ordering;
 use core::sync::atomic::AtomicU32;
 
-// Async communication with other threads
-static COUNT: AtomicU32 = AtomicU32::new(0); // For profiling: # of refreshes
-static VALUE: AtomicU32 = AtomicU32::new(0); // Value, 0-99, to display
-
 pub struct C5412Pins {
     pub p11: embassy_stm32::gpio::Output<'static, AnyPin>,
     pub p12: embassy_stm32::gpio::Output<'static, AnyPin>,
@@ -136,13 +132,16 @@ impl C5412Pins {
 }
 
 #[embassy_executor::task]
-pub async fn process(c5412pins_ref: &'static mut C5412Pins) {
+pub async fn process(c5412pins_ref: &'static mut C5412Pins,
+                     count_atomic: &'static AtomicU32, // Profiling
+                     value_atomic: &'static AtomicU32) // What value to display
+{
     const ON_TIME_MS : u64 = 2;
     const OFF_TIME_MS : u64 = 5;
     let mut count : u32 = 0; // For profiling the refresh rate. Just counts
     loop {
-        COUNT.store(count, Ordering::Relaxed);
-        let x : u32 = VALUE.load(Ordering::Relaxed); // What to display
+        count_atomic.store(count, Ordering::Relaxed);
+        let x : u32 = value_atomic.load(Ordering::Relaxed); // What to display
 
         c5412pins_ref.all_off();
         Timer::after_millis(OFF_TIME_MS).await;
@@ -158,14 +157,4 @@ pub async fn process(c5412pins_ref: &'static mut C5412Pins) {
 
         count+=1;
     }
-}
-
-// API: call `set_value` with your two digit integer, 0-99
-pub fn set_value(value: u32) {
-    VALUE.store(value, Ordering::Relaxed);
-}
-
-// For profiling
-pub fn get_count() -> u32 {
-    COUNT.load(Ordering::Relaxed)
 }
